Remove full paths in package.json
Thank you for this example, it's exactly what I was looking for! Just a minor adjustment, a really nice (and seemingly not well-known) thing about npm is that you don't need to put the full path to a binary within "scripts". And yes, this will work even if you do not install with the -g flag.